requirements.txt
Django==3.0.6
django-cors-headers==3.3.0
django-cors-middleware==1.5.0
django-filter==2.2.0
django-rest==0.0.1
djangorestframework==3.11.0
pytz==2020.1
psycopg2==2.8.5
djangorestframework==3.11.0
djangorestframework-jwt==1.8.0